+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.5
+
+### Added
+- Make loading lua script play better with Redis clusters.
+- Handle failures loading lua script on Redis server to prevent infinite loop.
+
 ## 1.0.4
+
+### Fixed
 - Fix wait_time method to match the documentation from [bc-swoop](https://github.com/bc-swoop)
 
 ## 1.0.3